pyproject.toml
[tool.poetry]
version = "0.6.8"
name = "cython-extensions-sc2"
description = ""
authors = ["Tom Kerr <tomkerrdev@gmail.com>"]
readme = "README.md"
packages = [{include = "cython_extensions"}]
 include = [
   { path = "cython_extensions/*.so", format = "wheel" },
   { path = "cython_extensions/*.pyd", format = "wheel" }
 ]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
burnysc2 = ">=7.0.0"
jupyterlab = "^4.4.6"

[tool.poetry.group.dev.dependencies]
cython = "^3.0.8"
ipykernel = "^6.29.0"
line-profiler = "^4.1.2"
matplotlib = "^3.8.2"
notebook = "^7.3.3"
numpy = "^2.0.0"
black = "^24.1.1"
isort = "^5.13.2"
ipympl = "^0.9.3"
ipywidgets = "^8.1.7"

[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.21.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.0.6"
mkdocstrings-python = "^1.13.0"

[tool.poetry.group.semver.dependencies]
python-semantic-release = "9.21.0"

[build-system]
requires = ["poetry-core", "Cython", "numpy", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.build]
generate-setup-file = false
script = 'build.py'

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider" # deactivating pytest caching.
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]

[tool.coverage.report]
exclude_lines = [
    'if __name__ == "__main__":'
]

[tool.isort]
profile = "black"

[tool.semantic_release]
assets = []
build_command_env = []
commit_message = "{version}\n\nAutomatically generated by python-semantic-release"
commit_parser = "angular"
logging_use_named_masks = false
major_on_zero = true
allow_zero_version = true
no_git_verify = false
tag_format = "v{version}"
version_variables = [
  "pyproject.toml:version",
  "cython_extensions/__init__.py:__version__"
]

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "rc"
prerelease = false

[tool.semantic_release.changelog]
exclude_commit_patterns = []
mode = "init"
insertion_flag = "<!-- version list -->"
template_dir = "templates"

[tool.semantic_release.changelog.default_templates]
changelog_file = "CHANGELOG.md"
output_format = "md"
mask_initial_release = false

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
trim_blocks = false
lstrip_blocks = false
newline_sequence = "\n"
keep_trailing_newline = false
extensions = []
autoescape = false

[tool.semantic_release.commit_author]
env = "GIT_COMMIT_AUTHOR"
default = "semantic-release <semantic-release>"

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]
other_allowed_tags = ["build", "chore", "ci", "docs", "style", "refactor", "test"]
allowed_tags = ["feat", "fix", "perf", "build", "chore", "ci", "docs", "style", "refactor", "test"]
default_bump_level = 0

[tool.semantic_release.remote]
name = "origin"
type = "github"
ignore_token_for_push = false
insecure = false

[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true
